[ALPS04923984] MFB: pmqos dsdn vcore k
[Detail] change vcore

[Solution]

MTK-Commit-Id: b12f3db0c18959ccf4018649c645cdb97e4b7be8

CR-Id: ALPS04923984
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/cameraisp/mfb/isp_6s/camera_mfb.h
@@ -313,7 +313,7 @@ struct MFB_MSSConfig {
 	struct mss_dma dmas[TPIPE_NUM_PER_FRAME];
 	unsigned int update_dma_en[TPIPE_NUM_PER_FRAME];
 	unsigned int tpipe_used;
-	unsigned int qos;
+	unsigned long qos;
 };
 
 struct MFB_MSFConfig {
@@ -325,7 +325,7 @@ struct MFB_MSFConfig {
 	unsigned int MFBDMT_TDRI_OFST[TPIPE_NUM_PER_FRAME];
 	unsigned int MFBDMT_TDRI_XSIZE[TPIPE_NUM_PER_FRAME];
 	unsigned int tpipe_used;
-	unsigned int qos;
+	unsigned long qos;
 };
 
 /******************************************************************************
